added ability to delete bodies
I ended up using a seperate package, because the normal SwiftUI elements didn't seem to be working. I tried using a custom SwiftUI implementation of UISheetPresentationController, but that didn't update with state, and then I tried using the built in popover(), but that was also didn't provide ideal results and wasn't as customisable as I would have hoped. In the end, this package seemed to work well for me :D